+++ src/mmda/recipes/__init__.py
@@ -1,0 +1,5 @@
+from mmda.recipes.core_recipe import CoreRecipe
+
+__all__ = [
+    'CoreRecipe'
+]
